Refactor files into two seperate conditions
