+++ src/db/table/select/mod.rs
@@ -1,4 +1,5 @@
 pub mod where_stack;
+pub mod where_condition;
 pub mod limit_clause;
 pub mod order_by_clause;
 use crate::db::table::{Table, Value};
